sql: accept spaces inside array type modifiers
PostgreSQL accepts NUMERIC(10, 2); allow optional spaces inside the
parenthesised digit list of a type modifier without widening what the
parentheses may contain. Move the mock-server MySQL TLS test into its own
non-Docker test file and tighten stderr assertions in the SQL fixtures.
