+++ src/js/internal/sql/postgres.ts
@@ -216,7 +216,11 @@ function getArrayType(typeNameOrID: number | ArrayType | undefined = undefined):
     // dots, each optionally followed by a `(digits[,digits])` modifier.
     // Parentheses may only wrap digit lists so the value can never close the
     // enclosing expression and break out of the `$N::${type}[]` cast.
-    if (!/^[A-Z_][A-Z0-9_]*(\([0-9]+(,[0-9]+)*\))?([ .][A-Z_][A-Z0-9_]*(\([0-9]+(,[0-9]+)*\))?)*$/.test(type)) {
+    if (
+      !/^[A-Z_][A-Z0-9_]*(\( *[0-9]+( *, *[0-9]+)* *\))?([ .][A-Z_][A-Z0-9_]*(\( *[0-9]+( *, *[0-9]+)* *\))?)*$/.test(
+        type,
+      )
+    ) {
       throw $ERR_INVALID_ARG_VALUE("type", typeNameOrID, "must be a valid PostgreSQL type name");
     }
     return type;
